OK - Pretty sure A* is working correctly, it's just the heuristic...
Because the gScores are imaginary and the hScore is physical distance, it's not coinciding with the cost (gScore).

if all distances were set (eg 10 orthogonal, 14 diagonal) and the distance calculated by (#nodes vert + #nodes horizontal) * 10, it'd give a realistic estimate.

So need to listen to the second A* lecture & tutorial, but then move onto Agents, etc for Monday.